[air] Allow herd tiling for degenerate herd sizes
Currently, after tiling, we have an intermediate point where a herd is
represented by a 2D scf.parallel loop.  However, if tiled with a size that
results in a herd dimension of one, then the parallel loops can
get optimized, resulting in a loop that looks different.  This patch
moves the optimization of loops until after an air.herd_launch op is
created, simplifying the required matching.
This patch also fixes the functionality of the -herd-sizes parameter
so that it takes up to three sizes.  The first two are the spatial sizes,
the third is a temporal size executed on the core.