fix(ATenToXTen.td): handle the case where bias is none
the torch linear operator can have the bias set to none, here we handle the
case by simply passing the value along. As our torch->onnx flow will
handle inserting an empty bias tensor. If we get torch mlir from say
torchscript, and it contains a none bias then we pass it along.

In XtenToLinAlg the empty bias is handled and an empty bias tensor is
generated. Following the convention of xten we simply pass along.