feat(Quantization): add operation definitions
Add definitions for quantize and dequantize operations that can
convert float tensors from and to integer tensors.

Note, the types supported are tied highly to the ONNX representation
and limited to zero_point == 0 and scale in power of two values.

Signless integers are not supported as quantization specifies the
signedness of the tensor.