feat(XTenNNToTOSA): conversion pass
decompose QDQ operations in XTenNN to TOSA

Support only signed integer operations at the moment, as TOSA can only
accept signed integers (represented as signless) for most of the
operations.

Take the following XTenNN operations:
```
%0 = xten_nn.quantize(%arg0 : tensor<1xf32>) {shift = -5 : si32} -> tensor<1xsi8>
%1 = xten_nn.dequantize(%0 : tensor<1xsi8>) {shift = -5 : si32} -> tensor<1xf32>
```
and convert them to:
```
%0 = "tosa.const"(){value = dense<3.2000000e+01> : tensor<f32>} : () -> tensor<1xf32>
%1 = "tosa.const"(){value = dense<3.1250000e-02>: tensor<f32>} : () -> tensor<1xf32>
%2 = "tosa.mul"(%arg0, %0) {shift = 0: i32} :(tensor<1xf32>, tensor<1xf32>) -> tensor<1xf32>
%3 = "tosa.cast"(%2) : (tensor<1xf32>) -> tensor<1xi8>
%4 = "tosa.cast"(%3) : (tensor<1xi8>) -> tensor<1xf32>
%5 = "tosa.mul"(%4, %1) {shift = 0: i32} :(tensor<1xf32>, tensor<1xf32>) -> tensor<1xf32>
```
Clamps are used when a signed bit width for the QDQ operations are used
that do not match i8, i16 or i32 directly. However, in the above case
the clamp is folded away as the clamping matches the i8 type exactly.

In the future, unsigned support can be added provided TOSA is updated to
support ui<?> tensors.